refactor: 구조개선
