+++ server/src/routes/auth.ts
