+++ server/src/routes/hscode/:code.ts
@@ -1,8 +1,8 @@
 import express, { Request, Response, NextFunction } from 'express';
-import verifyToken from '../../../middlewares/verifyToken';
+import verifyToken from '../../middlewares/verifyToken';
 import axios from 'axios';
-import logger from "../../../utils/logger";
-import HttpException from "../../../exceptions/HttpException";
+import logger from "../../utils/logger";
+import HttpException from "../../exceptions/HttpException";
 
 
 const router = express.Router();
@@ -15,7 +15,6 @@ router.get('/:code', verifyToken, async (req: Request, res: Response, next: Next
         return res.status(proxy_response.status).json(proxy_response);
     } catch (error) {
         if (axios.isAxiosError(error)) {
-            console.log(error)
             return res.status(400).json({ 'status': 400, message: 'Bad request' });
         } else {
             return next(error as HttpException);
