+++ server/src/routes/requests/index.ts
@@ -1,8 +1,8 @@
 import express, { Request, Response, NextFunction } from 'express';
-import { supabase } from '../../../utils/supabase';
-import verifyToken from "../../../middlewares/verifyToken";
-import insertGoods from "../../../middlewares/insertGoods";
-import { isoToUnix, unixToIso } from "../../../middlewares/timeConvert";
+import { supabase } from '../../utils/supabase';
+import verifyToken from "../../middlewares/verifyToken";
+import insertGoods from "../../middlewares/insertGoods";
+import { isoToUnix, unixToIso } from "../../middlewares/timeConvert";
 
 
 const router = express.Router();
